requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
PyPDF2
chromadb
langchain_google_genai
langchain-community



